requirements.txt
# Core
python-dotenv==1.0.0
pyyaml==6.0.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Database
duckdb==0.9.2
sqlalchemy==2.0.23

# API clients
requests==2.31.0
urllib3==2.1.0

# Web scraping / Parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Streamlit & Visualization
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0
matplotlib==3.8.2
seaborn==0.13.0

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Optional: Toornament API (if official SDK exists)
# toornament-sdk==x.x.x
